Provide ability to set the relay type to different states
The different states can be set globally for all newly created batons orchanged on the fly for specifc ones. This can be helpful for users to access the type in the metadata of the baton and perform specific actions (or lack thereof) on batons in that state (i.e. exclude batons from a search and load if they are not of type competition). The possiblities for this control is left up to the user at this time and for the most part the functions will perform the same.

Future behaviour of the package could possiblly allow for functions to operate specific to the baton type. Likely this would be done by creating functions as generics and each baton having a subclass and method based on that type. However, given the flexibility afforded to how batons are passed in-memory and connected to YAML, this is quite difficult to predict what users will do. It may be that they want to ignore all batons, but still want convenience of creating bundles, which sometimes still requires a baton. As such, this is not planned as a feature in the short term. Some options would be to have behaviour consist of, crashing out of functions, returning dummy batons, or proceeding as normal.

A few other minor typos were also included in this commit.