Add new class for previewing batons with restricted functionality.
Mainly made so it is easier to plot and summarise a baton without having to 'grab' it as in a normal workflow. Designed to only be an interactive type of usage. It is recommended to use other functions that also work by peeking into the YAML file directly instead of having to load. The loading is only really required for when R needs to know the class for method dispatch, which YAML file cant provide without loading.